Python/Bots/Ben2.0/requirements.txt
speechrecognition
pyttsx3
pywhatkit
pygame
pyautogui
sounddevice
soundfile
numpy
torch
transformers
diffusers
resemblyzer
openai-whisper
pillow

